Fix potential bugs with signed variables

+++ sources/funconv.c
@@ -234,8 +234,8 @@ void calc_lookup_args(void)
   if (fcs.ta_clt >= ((int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]))) break;
 
  //lookup table works from clt_grid_points[0] and upper
- if (fcs.ta_i < 0)  {fcs.ta_i = 0; fcs.ta_clt = PGM_GET_WORD(&fw_data.exdata.clt_grid_points[0]);}
- if (fcs.ta_clt > ((int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[F_TMP_POINTS-1]))) fcs.ta_clt = PGM_GET_WORD(&fw_data.exdata.clt_grid_points[F_TMP_POINTS-1]);
+ if (fcs.ta_i < 0)  {fcs.ta_i = 0; fcs.ta_clt = (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[0]);}
+ if (fcs.ta_clt > ((int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[F_TMP_POINTS-1]))) fcs.ta_clt = (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[F_TMP_POINTS-1]);
  fcs.ta_i1 = fcs.ta_i + 1;
 
 #if !defined(SECU3T) && defined(MCP3204)
@@ -248,8 +248,8 @@ void calc_lookup_args(void)
   if (fcs.ga_grt >= ((int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ga_i]))) break;
 
  //lookup table works from clt_grid_points[0] and upper
- if (fcs.ga_i < 0)  {fcs.ga_i = 0; fcs.ga_grt = PGM_GET_WORD(&fw_data.exdata.clt_grid_points[0]);}
- if (fcs.ga_grt > ((int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[F_TMP_POINTS-1]))) fcs.ga_grt = PGM_GET_WORD(&fw_data.exdata.clt_grid_points[F_TMP_POINTS-1]);
+ if (fcs.ga_i < 0)  {fcs.ga_i = 0; fcs.ga_grt = (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[0]);}
+ if (fcs.ga_grt > ((int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[F_TMP_POINTS-1]))) fcs.ga_grt = (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[F_TMP_POINTS-1]);
  fcs.ga_i1 = fcs.ga_i + 1;
 #endif
 
@@ -309,10 +309,10 @@ int16_t coolant_function(uint8_t mode)
 
  if (mode) //work mode
   return simple_interpolation(fcs.ta_clt, _GB(f_tmp[fcs.ta_i]), _GB(f_tmp[fcs.ta_i1]),
-        PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16);
+        (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16);
  else //idling mode
   return simple_interpolation(fcs.ta_clt, _GB(f_tmp_idl[fcs.ta_i]), _GB(f_tmp_idl[fcs.ta_i1]),
-        PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16);
+        (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16);
 }
 
 int16_t crkclt_function(void)
@@ -321,7 +321,7 @@ int16_t crkclt_function(void)
   return 0;   //no correction if CLT sensor is turned off
 
  return simple_interpolation(fcs.ta_clt, (int8_t)PGM_GET_BYTE(&fw_data.exdata.cts_crkcorr[fcs.ta_i]), (int8_t)PGM_GET_BYTE(&fw_data.exdata.cts_crkcorr[fcs.ta_i1]),
-        PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16);
+        (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16);
 }
 
 //Idling regulator
@@ -626,7 +626,7 @@ uint16_t inj_cranking_pw(void)
   return 1000;   //coolant temperature sensor is not enabled, default is 3.2mS
 
  return simple_interpolation(fcs.ta_clt, _GWU(inj_cranking[fcs.ta_i]), _GWU(inj_cranking[fcs.ta_i1]),  //<--values in table are unsigned
-        PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 1) /*>> 0*/;
+        (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 1) /*>> 0*/;
 }
 
 void calc_ve_afr(void)
@@ -840,7 +840,7 @@ uint8_t inj_iac_pos_lookup(prev_temp_t* p_pt, uint8_t mode)
 #endif
  //run/cranking
  return simple_interpolation(t, mode ? _GBU(inj_iac_run_pos[i]) : _GBU(inj_iac_crank_pos[i]), mode ? _GBU(inj_iac_run_pos[i1]) : _GBU(inj_iac_crank_pos[i1]),  //<--values in table are unsigned
-  PGM_GET_WORD(&fw_data.exdata.clt_grid_points[i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[i]), 16) >> 4;
+  (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[i]), 16) >> 4;
 }
 #endif
 
@@ -851,7 +851,7 @@ uint8_t inj_aftstr_en(void)
   return 0;   //coolant temperature sensor is not enabled (or not installed), no afterstart enrichment
 
  return simple_interpolation(fcs.ta_clt, _GBU(inj_aftstr[fcs.ta_i]), _GBU(inj_aftstr[fcs.ta_i1]),  //<--values in table are unsigned
-        PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4;
+        (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4;
 }
 
 uint8_t inj_warmup_en(void)
@@ -860,7 +860,7 @@ uint8_t inj_warmup_en(void)
   return 128;   //coolant temperature sensor is not enabled (or not installed), no warmup enrichment
 
  return simple_interpolation(fcs.ta_clt, _GBU(inj_warmup[fcs.ta_i]), _GBU(inj_warmup[fcs.ta_i1]),  //<--values in table are unsigned
-        PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4;
+        (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4;
 }
 
 int16_t inj_ae_tps_lookup(int16_t tpsdot)
@@ -1024,7 +1024,7 @@ uint16_t inj_idling_rpm(void)
   return 900;   //coolant temperature sensor is not enabled (or not installed)
 
  return (simple_interpolation(fcs.ta_clt, _GBU(inj_target_rpm[fcs.ta_i]), _GBU(inj_target_rpm[fcs.ta_i1]),  //<--values in table are unsigned
-         PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4) * 10;
+         (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4) * 10;
 }
 #endif
 
@@ -1407,7 +1407,7 @@ uint16_t cranking_thrd_rpm(void)
   return d.param.starter_off;   //coolant temperature sensor is not enabled (or not installed), use simple constant
 
  return (simple_interpolation(fcs.ta_clt, PGM_GET_BYTE(&fw_data.exdata.cranking_thrd[fcs.ta_i]), PGM_GET_BYTE(&fw_data.exdata.cranking_thrd[fcs.ta_i1]),  //<--values in table are unsigned
-         PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4) * 10;
+         (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4) * 10;
 }
 
 uint16_t cranking_thrd_tmr(void)
@@ -1416,7 +1416,7 @@ uint16_t cranking_thrd_tmr(void)
   return PGM_GET_BYTE(&fw_data.exdata.stbl_str_cnt);   //coolant temperature sensor is not enabled (or not installed), use simple constant
 
  return (simple_interpolation(fcs.ta_clt, PGM_GET_BYTE(&fw_data.exdata.cranking_time[fcs.ta_i]), PGM_GET_BYTE(&fw_data.exdata.cranking_time[fcs.ta_i1]),  //<--values in table are unsigned
-         PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4) * 10;
+         (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4) * 10;
 }
 
 uint16_t smapaban_thrd_rpm(void)
@@ -1425,7 +1425,7 @@ uint16_t smapaban_thrd_rpm(void)
   return d.param.smap_abandon;   //coolant temperature sensor is not enabled (or not installed). use simple constant
 
  return (simple_interpolation(fcs.ta_clt, PGM_GET_BYTE(&fw_data.exdata.smapaban_thrd[fcs.ta_i]), PGM_GET_BYTE(&fw_data.exdata.smapaban_thrd[fcs.ta_i1]),  //<--values in table are unsigned
-         PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4) * 10;
+         (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4) * 10;
 }
 
 #ifdef _PLATFORM_M1284_
@@ -1501,13 +1501,13 @@ int16_t split_function(void)
 uint8_t grheat_pwm_duty(void)
 {
  return simple_interpolation(fcs.ga_grt, PGM_GET_BYTE(&fw_data.exdata.grheat_duty[fcs.ga_i]), PGM_GET_BYTE(&fw_data.exdata.grheat_duty[fcs.ga_i1]),
-        PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ga_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ga_i]), 16) >> 4;
+        (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ga_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ga_i]), 16) >> 4;
 }
 
 uint16_t grv_delay(void)
 {
  return simple_interpolation(fcs.ga_grt, PGM_GET_WORD(&fw_data.exdata.grv_delay[fcs.ga_i]), PGM_GET_WORD(&fw_data.exdata.grv_delay[fcs.ga_i1]),
-        PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ga_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ga_i]), 2) >> 1;
+        (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ga_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ga_i]), 2) >> 1;
 }
 #endif
 
@@ -1539,7 +1539,7 @@ uint16_t aftstr_strokes(uint8_t mode)
  { //gas
   if (0==d.param.inj_aftstr_strokes1)
    return simple_interpolation(fcs.ta_clt, PGM_GET_BYTE(&fw_data.exdata.inj_aftstr_strk1[fcs.ta_i]), PGM_GET_BYTE(&fw_data.exdata.inj_aftstr_strk1[fcs.ta_i1]),  //<--values in table are unsigned
-         PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4;
+         (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4;
   else
    return ((uint16_t)d.param.inj_aftstr_strokes1) << 2;
  }
@@ -1547,7 +1547,7 @@ uint16_t aftstr_strokes(uint8_t mode)
  { //petrol
   if (0==d.param.inj_aftstr_strokes)
    return simple_interpolation(fcs.ta_clt, PGM_GET_BYTE(&fw_data.exdata.inj_aftstr_strk0[fcs.ta_i]), PGM_GET_BYTE(&fw_data.exdata.inj_aftstr_strk0[fcs.ta_i1]),  //<--values in table are unsigned
-         PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4;
+         (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_points[fcs.ta_i]), (int16_t)PGM_GET_WORD(&fw_data.exdata.clt_grid_sizes[fcs.ta_i]), 16) >> 4;
   else
    return ((uint16_t)d.param.inj_aftstr_strokes) << 2;
  }
